Make spawn.json schemas module-private, restoring z.infer for SpawnConfig
The exported schemas required hand-declared interfaces because
isolatedDeclarations forbids inferred types on exports, leaving the
runtime validators and compile-time types to be kept in sync manually.
Nothing outside SpawnEngine uses the schemas, so they move into the
engine as private consts where z.infer works again. Only PropertySchema
remains public — it is part of the PropertyResolver contract — and the
private schema is annotated against it.
