+++ kinotic-js/workspace/packages/spawn/src/index.ts
@@ -1,6 +1,6 @@
 export * from './api/PropertyResolver'
+export * from './api/PropertySchema'
 export * from './api/RenderSpawnOptions'
-export * from './api/SpawnConfig'
 export * from './api/SpawnEngine'
 export * from './api/SpawnRenderResult'
 export * from './api/SpawnTree'
